requirements.txt
torch~=2.4.0
numpy~=2.0.1
opencv-python~=4.10.0.84
moviepy~=1.0.3
colorama~=0.4.6
stribor
torchdiffeq
scipy~=1.14.0
tqdm~=4.66.4
pandas~=2.2.2
matplotlib~=3.9.1
scikit-learn~=1.5.1
torchvision~=0.19.0
timm~=1.0.7
pillow~=10.4.0